requirements.txt
tensorflow
keras
python-telegram-bot